Fixed failing test in record_test.
TestRecordGetFieldSchema was failing.  The issue is that it was
passing an invalid name to GetFieldSchema, which then tried create
a name record out of it and failed, causing a nil pointer later
on, as the error created when attempting to create the record
schema wasn't properly handled.

Fixed the incomplete error handling and added a test for that while
I was at it.