Start fixing namespace handling.
This starts to fix namespace handling such that it's more consistent
with the Java implementation.  fullnames, that is, names prepended
with a namespace, are now allowed as record names.

There was already a (failing) test case for this, which now passes.

There are some other inconsistencies between the Java version and
the spec that are not yet handled, the biggest one is that we're
not currently validating names for fields in records.